Split metadata-handling classes out of 'utils' into new 'metadata' module.
